services/actor-type-svc/pyproject.toml
[project]
name = "actor-type-svc"
version = "2.0.0"
description = "UC11 Actor Type Service — gRPC-Microservice fuer Akteur-Typ-Verteilungsanalyse"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "TI-Radar Team" }]

dependencies = [
    "grpcio>=1.68.0", "grpcio-tools>=1.68.0", "grpcio-health-checking>=1.68.0",
    "grpcio-reflection>=1.68.0", "protobuf>=6.31", "asyncpg>=0.30.0", "pydantic-settings>=2.6.0",
    "prometheus-client>=0.21.0", "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "pytest-asyncio>=0.24.0", "pytest-cov>=6.0.0", "ruff>=0.8.0", "mypy>=1.13.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
